:bug: Fix int field in camunda forms (#1200)

+++ frontend/zac-ui/libs/features/zaak-detail/src/lib/actions/keten-processen/configuration-components/dynamic-form/dynamic-form.component.ts
@@ -103,7 +103,7 @@ export class DynamicFormComponent implements OnChanges {
         case 'date':
           value = this.datePipe.transform(value, "yyyy-MM-dd hh:mm:ss");
           break;
-        case 'int': value = parseInt(value, 10); break;
+        case 'long': value = parseInt(value, 10); break;
       }
       formData[control] = value;
     })
